Rename mock to util function
