+++ app/steps/pay/card-payment-status/index.test.js
@@ -7,7 +7,7 @@ const idamMock = require('test/mocks/idam');
 
 const { expect, sinon } = require('test/util/chai');
 const server = require('app');
-const featureTogglesMock = require('test/mocks/featureToggles');
+const featureToggleConfig = require('test/util/featureToggles');
 const { testRedirect, testCustom } = require('test/util/assertions');
 const serviceToken = require('app/services/serviceToken');
 const payment = require('app/services/payment');
@@ -120,26 +120,26 @@ describe(modulePath, () => {
         // Arrange.
         const userCookie = ['__auth-token=auth.token'];
         // Act.
-        const featureMock = featureTogglesMock
+        const featureToggle = featureToggleConfig
           .when('idam', true, testCustom, agent, underTest, userCookie, () => {
             // Assert.
             expect(query.calledOnce).to.equal(true);
             expect(query.args[0][0]).to.eql({ id: 1, bearerToken: 'auth.token' });
           });
-        featureMock(done);
+        featureToggle(done);
       });
     });
 
     context('Idam is turned OFF', () => {
       it('uses a fake user for the mocks', done => {
         // Act.
-        const featureMock = featureTogglesMock
+        const featureToggle = featureToggleConfig
           .when('idam', false, testCustom, agent, underTest, [], () => {
             // Assert.
             expect(query.calledOnce).to.equal(true);
             expect(query.args[0][0]).to.eql({});
           });
-        featureMock(done);
+        featureToggle(done);
       });
     });
