+++ app/steps/pay/pay-online-only/index.test.js
@@ -3,7 +3,7 @@ const request = require('supertest');
 const server = require('app');
 const idamMock = require('test/mocks/idam');
 const { testContent, testCustom, getSession } = require('test/util/assertions');
-const featureTogglesMock = require('test/mocks/featureToggles');
+const featureToggleConfig = require('test/util/featureToggles');
 const applicationFeeMiddleware = require('app/middleware/updateApplicationFeeMiddleware');
 const getBaseUrl = require('app/core/utils/baseUrl');
 const { expect, sinon } = require('test/util/chai');
@@ -183,26 +183,26 @@ describe(modulePath, () => {
       context('Idam is turned ON', () => {
         it('uses the token of the logged in user', done => {
           // Act.
-          const featureMock = featureTogglesMock
+          const featureToggle = featureToggleConfig
             .when('idam', true, testCustom, agent, underTest, cookies, () => {
               // Assert.
               expect(create.calledOnce).to.equal(true);
               expect(create.args[0][0]).to.eql({ id: 1, bearerToken: 'auth.token' });
             }, 'post');
-          featureMock(done);
+          featureToggle(done);
         });
       });
 
       context('Idam is turned OFF', () => {
         it('uses a fake user for the mocks', done => {
           // Act.
-          const featureMock = featureTogglesMock
+          const featureToggle = featureToggleConfig
             .when('idam', false, testCustom, agent, underTest, [], () => {
               // Assert.
               expect(create.calledOnce).to.equal(true);
               expect(create.args[0][0]).to.eql({});
             }, 'post');
-          featureMock(done);
+          featureToggle(done);
         });
       });
